fix(BaseTransition): collect correct children with `Transition` slot
fix #1455
