+++ packages/runtime-core/src/components/BaseTransition.ts
@@ -8,7 +8,8 @@ import {
   Comment,
   isSameVNodeType,
   VNode,
-  VNodeArrayChildren
+  VNodeArrayChildren,
+  Fragment
 } from '../vnode'
 import { warn } from '../warning'
 import { isKeepAlive } from './KeepAlive'
@@ -135,7 +136,9 @@ const BaseTransitionImpl = {
     const state = useTransitionState()
 
     return () => {
-      const children = slots.default && slots.default()
+      const children = getTransitionRawChildren(
+        slots.default ? slots.default() : []
+      )
       if (!children || !children.length) {
         return
       }
@@ -416,4 +419,20 @@ export function setTransitionHooks(vnode: VNode, hooks: TransitionHooks) {
   } else {
     vnode.transition = hooks
   }
+}
+
+export function getTransitionRawChildren(children: VNode[]): VNode[] {
+  let ret: VNode[] = []
+  for (let i = 0; i < children.length; i++) {
+    const child = children[i]
+    // handle fragment children case, e.g. v-for
+    if (child.type === Fragment) {
+      ret = ret.concat(getTransitionRawChildren(child.children as VNode[]))
+    }
+    // comment placeholders should be skipped, e.g. v-if
+    else if (child.type !== Comment) {
+      ret.push(child)
+    }
+  }
+  return ret
 }
