+++ packages/runtime-dom/src/components/TransitionGroup.ts
@@ -9,7 +9,6 @@ import {
 } from './Transition'
 import {
   Fragment,
-  Comment,
   VNode,
   warn,
   resolveTransitionHooks,
@@ -22,6 +21,7 @@ import {
 } from '@vue/runtime-core'
 import { toRaw } from '@vue/reactivity'
 import { extend } from '@vue/shared'
+import { getTransitionRawChildren } from '../../../runtime-core/src/components/BaseTransition'
 
 interface Position {
   top: number
@@ -127,22 +127,6 @@ const TransitionGroupImpl = {
       return createVNode(tag, null, slotChildren)
     }
   }
-}
-
-function getTransitionRawChildren(children: VNode[]): VNode[] {
-  let ret: VNode[] = []
-  for (let i = 0; i < children.length; i++) {
-    const child = children[i]
-    // handle fragment children case, e.g. v-for
-    if (child.type === Fragment) {
-      ret = ret.concat(getTransitionRawChildren(child.children as VNode[]))
-    }
-    // comment placeholders should be skipped, e.g. v-if
-    else if (child.type !== Comment) {
-      ret.push(child)
-    }
-  }
-  return ret
 }
 
 // remove mode props as TransitionGroup doesn't support it
